requirements.txt
asgiref==3.5.2
Django==4.1
django-enum-choices==2.1.4
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
phonenumbers==8.12.54
pytz==2022.2.1
sqlparse==0.4.2
